pyproject.toml
[project]
name = "credit-card-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-mistralai>=1.1.1",
    "mistralai>=1.10.1",
    "pinecone>=8.0.0",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]
